Fixing bugs and adding tests such that every major component has a test
